guardar 5

+++ 2011112243/2011112243/Venta.cs
@@ -9,7 +9,7 @@ namespace _2011112243
    public class Venta
     {
         public String idVenta { get; set; }
-        public decimal Costo { get; set; }
+        public double Costo { get; set; }
         public Administrativo _Administrativo;
         public Servicio _Servicio;
         public Cliente _Cliente;
